# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Token {
  assetId: string;
  provider: string;
  mint: string;
  symbol: string;
  name: string;
  price: number;
  change24h: number;
}

export type RoundMode = typeof RoundMode[keyof typeof RoundMode];


export const RoundMode = {
  TEST: 'TEST',
  PRODUCTION: 'PRODUCTION',
} as const;

export interface Round {
  id: string;
  mode: RoundMode;
  opensAt: string;
  closesAt: string;
  resolvesAt: string;
  participantCount: number;
}

export type LeaderboardEntryDirection = typeof LeaderboardEntryDirection[keyof typeof LeaderboardEntryDirection];


export const LeaderboardEntryDirection = {
  LONG: 'LONG',
  SHORT: 'SHORT',
} as const;

export interface LeaderboardEntry {
  rank: number;
  wallet: string;
  tokenSymbol: string;
  direction: LeaderboardEntryDirection;
  returnPct: number;
  isMe: boolean;
}

export type SubmissionDirection = typeof SubmissionDirection[keyof typeof SubmissionDirection];


export const SubmissionDirection = {
  LONG: 'LONG',
  SHORT: 'SHORT',
} as const;

export interface Submission {
  id: string;
  wallet: string;
  roundId: string;
  assetId: string;
  tokenMint: string;
  tokenSymbol: string;
  direction: SubmissionDirection;
  entryPrice: number;
  entryTimestamp: string;
}

export interface CompetitionState {
  round: Round;
  tokens: Token[];
  leaderboard: LeaderboardEntry[];
  totalPlayers: number;
  mySubmission: Submission | null;
}

export type SubmissionInputDirection = typeof SubmissionInputDirection[keyof typeof SubmissionInputDirection];


export const SubmissionInputDirection = {
  LONG: 'LONG',
  SHORT: 'SHORT',
} as const;

export interface SubmissionInput {
  /** @minLength 1 */
  wallet: string;
  /** @minLength 1 */
  assetId: string;
  direction: SubmissionInputDirection;
}

export type HistoryEntryMode = typeof HistoryEntryMode[keyof typeof HistoryEntryMode];


export const HistoryEntryMode = {
  TEST: 'TEST',
  PRODUCTION: 'PRODUCTION',
} as const;

export type HistoryEntryDirection = typeof HistoryEntryDirection[keyof typeof HistoryEntryDirection];


export const HistoryEntryDirection = {
  LONG: 'LONG',
  SHORT: 'SHORT',
} as const;

export type HistoryEntryStatus = typeof HistoryEntryStatus[keyof typeof HistoryEntryStatus];


export const HistoryEntryStatus = {
  WON: 'WON',
  LOST: 'LOST',
  PENDING: 'PENDING',
} as const;

export interface HistoryEntry {
  roundId: string;
  mode: HistoryEntryMode;
  tokenSymbol: string;
  direction: HistoryEntryDirection;
  returnPct: number;
  status: HistoryEntryStatus;
  date: string;
}

export type AdminSettingsMode = typeof AdminSettingsMode[keyof typeof AdminSettingsMode];


export const AdminSettingsMode = {
  TEST: 'TEST',
  PRODUCTION: 'PRODUCTION',
} as const;

export interface AdminSettings {
  mode: AdminSettingsMode;
  tokens: Token[];
}

export type AdminSettingsInputMode = typeof AdminSettingsInputMode[keyof typeof AdminSettingsInputMode];


export const AdminSettingsInputMode = {
  TEST: 'TEST',
  PRODUCTION: 'PRODUCTION',
} as const;

export interface AdminSettingsInput {
  mode: AdminSettingsInputMode;
  /** @items.minLength 1 */
  tokenMints: string[];
}

export type SearchMarketsParams = {
/**
 * @minLength 2
 */
q: string;
};

export type GetMyHistoryParams = {
wallet: string;
};