# lib/api-zod/src/generated/types/token.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Token {
  assetId: string;
  provider: string;
  mint: string;
  symbol: string;
  name: string;
  price: number;
  change24h: number;
}